refactor: extract User and Permission classes
 Classes used to model User-permissions in IT
