+++ qa/integration-tests/src/test/java/io/camunda/it/auth/TenantAuthorizationIT.java
@@ -17,8 +17,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.client.CamundaClient;
 import io.camunda.it.utils.BrokerITInvocationProvider;
 import io.camunda.it.utils.CamundaClientTestFactory.Authenticated;
-import io.camunda.it.utils.CamundaClientTestFactory.Permissions;
-import io.camunda.it.utils.CamundaClientTestFactory.User;
+import io.camunda.qa.util.auth.Permissions;
+import io.camunda.qa.util.auth.User;
 import io.camunda.zeebe.util.Either;
 import io.camunda.zeebe.util.collection.Tuple;
 import java.io.IOException;
