+++ qa/integration-tests/src/test/java/io/camunda/it/auth/UserSearchingAuthorizationIT.java
@@ -17,8 +17,8 @@ import io.camunda.client.CamundaClient;
 import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.it.utils.BrokerITInvocationProvider;
 import io.camunda.it.utils.CamundaClientTestFactory.Authenticated;
-import io.camunda.it.utils.CamundaClientTestFactory.Permissions;
-import io.camunda.it.utils.CamundaClientTestFactory.User;
+import io.camunda.qa.util.auth.Permissions;
+import io.camunda.qa.util.auth.User;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
